[Fix,Refactor] Study 엔티티 위치 변경, 파일 업로드 API fix
