+++ src/main/java/com/example/wantudy/study/dto/StudyCategoryDto.java
@@ -1,7 +1,7 @@
 package com.example.wantudy.study.dto;
 
 import com.example.wantudy.study.domain.Category;
-import com.example.wantudy.study.domain.Study;
+import com.example.wantudy.study.Study;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
