+++ src/main/java/com/example/wantudy/study/dto/StudyDetailResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.wantudy.study.dto;
 
-import com.example.wantudy.study.domain.Study;
+import com.example.wantudy.study.Study;
 import com.example.wantudy.study.domain.StudyStatus;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,7 +8,6 @@ import lombok.Setter;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Date;
 import java.util.List;
 
 @Getter
