+++ src/main/java/com/example/wantudy/study/repository/StudyFileRepository.java
@@ -1,6 +1,6 @@
 package com.example.wantudy.study.repository;
 
-import com.example.wantudy.study.domain.Study;
+import com.example.wantudy.study.Study;
 import com.example.wantudy.study.domain.StudyFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
