+++ src/main/java/com/example/wantudy/study/repository/StudyRepository.java
@@ -1,7 +1,6 @@
 package com.example.wantudy.study.repository;
 
-import com.example.wantudy.study.domain.Category;
-import com.example.wantudy.study.domain.Study;
+import com.example.wantudy.study.Study;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
